Refactor: Separate resume components into Sidebar, MainContent, and ResumeDocument for better integration and maintainability
