+++ package.json
@@ -16,14 +16,14 @@
     "@mui/icons-material": "^5.16.1",
     "@mui/lab": "^5.0.0-alpha.172",
     "@mui/material": "^5.16.1",
-    "@react-pdf-viewer/core": "^3.12.0",
-    "@react-pdf-viewer/default-layout": "^3.12.0",
+    "@react-pdf-viewer/core": "^3.11.0",
+    "@react-pdf-viewer/default-layout": "^3.11.0",
     "@react-pdf/renderer": "^3.4.4",
     "@react-three/drei": "^9.108.3",
     "@react-three/fiber": "^8.16.8",
     "framer-motion": "^11.3.2",
     "framer-motion-3d": "^11.3.2",
-    "pdfjs-dist": "^3.4.120",
+    "pdfjs-dist": "^3.11.174",
     "react": "^18.2.0",
     "react-dom": "^18.3.1",
     "react-pdf": "^9.1.0",
