+++ src/components/PDFViewer.jsx
@@ -1,20 +1,16 @@
 import React from 'react';
-import { Worker } from '@react-pdf-viewer/core';
+import { Worker, Viewer } from '@react-pdf-viewer/core';
 import { defaultLayoutPlugin } from '@react-pdf-viewer/default-layout';
 import '@react-pdf-viewer/core/lib/styles/index.css';
 import '@react-pdf-viewer/default-layout/lib/styles/index.css';
-import { Viewer } from '@react-pdf-viewer/core';
 
-const PDFViewer = () => {
+const PDFViewer = ({ pdfBlobUrl }) => {
   const defaultLayoutPluginInstance = defaultLayoutPlugin();
 
   return (
     <div style={{ height: '750px' }}>
       <Worker workerUrl={`https://unpkg.com/pdfjs-dist@3.11.174/build/pdf.worker.min.js`}>
-        <Viewer
-          fileUrl="/path/to/your-file.pdf"
-          plugins={[defaultLayoutPluginInstance]}
-        />
+        <Viewer fileUrl={pdfBlobUrl} plugins={[defaultLayoutPluginInstance]} />
       </Worker>
     </div>
   );
